core-js-compat: Add more typescript types
