+++ packages/core-js-compat/get-modules-list-for-target-version.d.ts
@@ -1,0 +1,5 @@
+import type { ModuleName, TargetVersion } from "./shared";
+
+declare function getModulesListForTargetVersion(version: TargetVersion): readonly ModuleName[];
+
+export = getModulesListForTargetVersion;
