+++ packages/core-js-compat/shared.d.ts
@@ -1,0 +1,27 @@
+
+export type ModuleName = string;
+
+export type Target =
+  | 'android'
+  | 'bun'
+  | 'chrome'
+  | 'chrome-android'
+  | 'deno'
+  | 'edge'
+  | 'electron'
+  | 'firefox'
+  | 'firefox-android'
+  | 'hermes'
+  | 'ie'
+  | 'ios'
+  | 'node'
+  | 'opera'
+  | 'opera-android'
+  | 'phantom'
+  | 'quest'
+  | 'react-native'
+  | 'rhino'
+  | 'safari'
+  | 'samsung';
+
+export type TargetVersion = string;
